+++ ecoli/library/data_predicates.py
