+++ ecoli/processes/protein_degradation.py
@@ -15,11 +15,7 @@ import numpy as np
 from vivarium.core.process import Process
 from vivarium.core.composition import simulate_process_in_experiment
 
-from wholecell.utils.constants import REQUEST_PRIORITY_DEGRADATION
-from wholecell.utils import units
-
-from functools import reduce
-from utils.data_predicates import monotonically_increasing, monotonically_decreasing, all_nonnegative, approx_poisson
+from ecoli.library.data_predicates import monotonically_increasing, monotonically_decreasing, all_nonnegative
 
 
 class ProteinDegradation(Process):
